+++ Source/GroupControl.cpp
@@ -90,7 +90,7 @@ void GroupControl::SaveLayout(ofxJSONElement& moduleInfo)
 {
    IDrawableModule::SaveLayout(moduleInfo);
    
-   moduleInfo["uicontrols"].resize(mControlCables.size()-1);
+   moduleInfo["uicontrols"].resize((unsigned int)mControlCables.size()-1);
    for (int i=0; i<mControlCables.size()-1; ++i)
    {
       string controlName = "";
