+++ Source/Selector.cpp
@@ -114,7 +114,7 @@ void Selector::SaveLayout(ofxJSONElement& moduleInfo)
 {
    IDrawableModule::SaveLayout(moduleInfo);
    
-   moduleInfo["uicontrols"].resize(mControlCables.size()-1);
+   moduleInfo["uicontrols"].resize((unsigned int)mControlCables.size()-1);
    for (int i=0; i<mControlCables.size()-1; ++i)
    {
       string controlName = "";
